add support for generating release note from branch
Previously, the generate command would only work for generating a
release note using the latest release created from the repository. This
was changed so that release branches that create older releases can
still generate release notes from the release branch. For example, if
the latest release is v6.0.0 and I have a v5.5.x release branch with the
latest release on that branch being v5.5.5, I can generate a release
note for that release branch for my next release of v5.5.6.

It achieves this by grabbing all the last 50 releases made to the
repository and then getting the commit SHAs that are tagged for those
releases. It then goes through all the commits on the branch and tries
to match it to one of the 50 commit release SHAs. If it finds one, it
will fetch all the prs made after that commit.

Also changed some formatting of the release note template so that it
will not bold all the text for each pr.

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>