add flag to support last commit SHA
if this flag is given, it will generate a release note using all prs
merged between the last release made on the branch until this last
commit SHA that is passed in through the flag. If it is not given, it
will generate a release note using all prs from last release until the
latest commit on the branch.

Signed-off-by: Clara Fu <cfu@pivotal.io>