rename no-impact to misc
The old `release/no-impact` label is too narrowly scoped for how we use
it. We want to include prs that make changes to the internal project
workflow, which do not really belong within a "No Impact" section.
Therefore, I renamed the section to be "Miscellaneous" so that it will
be more generic.

Signed-off-by: Clara Fu <cfu@pivotal.io>