wip: fetch commits from latest minor/major release
We want to start fetching commits from the last major/minor release and
ignore patch releases because in cases where we want to release v6.6.0,
if we have previously released v6.5.1 and v6.5.0, any commit merged
between the time that v6.5.1 and v6.5.0 was released will be ignored.
This is because we just grab all commits made after the latest release
(which was v6.5.1). So to fix this, we want to ignore any patch releases
and only start from any major/minor releases. But after this, we will
want to ignore any prs that were merged in to any patch releases so that
we won't mention them twice.

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>