add logic around when to start fetching prs for generate
if the version to release is a major/minor, we want to start fetching
prs from the last major or minor release. If the version to release is a
patch, we want to start fetching prs from the last release regardless of
major/minor/patch. As a result, major/minor release notes will include
prs that were released in previous patches.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Clara Fu <cfu@pivotal.io>