allow ignoring releases by regular expression
6.7.0 was a weird release since there was a lot of new behaviour added
to master between 6.6.0 and 6.7.0, but 6.7.0 received very few of
those changes. However, we have merged 6.7.x back into master, so the
tags are clearly present on master.  In this unusual case, we want to
ignore the 6.7.x release, since otherwise the PRs merged pre-6.7.0 won't
be included.

This commit adds the flag `--ignore-release-regex` which takes a regular
expression of release names to ignore.

e.g.
```
--ignore-release-regex '^v6\.7\.\d+'
```

matches all 6.7.x releases

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>