reorder the label precedence
We changed the order of the labels in terms of its precedence as an
experiment to see if it makes more sense to have prs that are labelled
with bug or enhancement to always be prioritized. But I think this
doesn't make sense in the actual way that people apply the labels. For
example, a very common use case is that when there is a bug fix for an
unreleased feature in concourse, we will apply the bug label and the
misc label because it does not affect any released versions of
concourse.

Signed-off-by: Clara Fu <fclara@vmware.com>